# backend-go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + MySQL + Redis 的高性能预测系统后端 API，支持体育比赛预测、用户投票、排行榜和实时更新功能\n\n## 功能特性\n- 🏆 体育比赛预测系统\n- 👥 用户注册和认证\n- 🗳️ 预测投票功能\n- 📊 实时排行榜\n- ⚡ WebSocket 实时通信\n- 🚀 高性能缓存策略\n\n## 技术栈\n- **后端**: Go + Gin + GORM\n- **数据库**: MySQL 8.0\n- **缓存**: Redis 6.0\n- **认证**: JWT\n\n## API 版本管理\n当前版本: v1.0，支持向后兼容",
        "title": "预测系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support Team",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/matches": {
            "get": {
                "description": "获取比赛列表，支持过滤和分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "获取比赛列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "赛事类型",
                        "name": "tournament",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "比赛状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "限制数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/match.Match"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的比赛",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "创建比赛",
                "parameters": [
                    {
                        "description": "比赛信息",
                        "name": "match",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/match.CreateMatchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/match.Match"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/matches/finished": {
            "get": {
                "description": "获取已结束的比赛列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "获取已结束的比赛",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "限制数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/match.Match"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/matches/upcoming": {
            "get": {
                "description": "获取即将开始的比赛列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "获取即将开始的比赛",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/match.Match"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/matches/{id}": {
            "get": {
                "description": "根据ID获取比赛详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "获取比赛详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "比赛ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/match.Match"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新比赛信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "更新比赛信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "比赛ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "match",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/match.UpdateMatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/match.Match"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/matches/{id}/cancel": {
            "post": {
                "description": "取消比赛",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "取消比赛",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "比赛ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/matches/{id}/result": {
            "post": {
                "description": "设置比赛结果并结束比赛",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "设置比赛结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "比赛ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "比赛结果",
                        "name": "result",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/match.SetResultRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/matches/{id}/start": {
            "post": {
                "description": "将比赛状态设置为进行中",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "开始比赛",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "比赛ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "match.CreateMatchRequest": {
            "type": "object",
            "required": [
                "start_time",
                "team_a",
                "team_b",
                "tournament"
            ],
            "properties": {
                "start_time": {
                    "type": "string"
                },
                "team_a": {
                    "type": "string",
                    "maxLength": 100
                },
                "team_b": {
                    "type": "string",
                    "maxLength": 100
                },
                "tournament": {
                    "$ref": "#/definitions/match.Tournament"
                }
            }
        },
        "match.Match": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Record creation timestamp",
                    "type": "string"
                },
                "id": {
                    "description": "Unique identifier",
                    "type": "integer"
                },
                "predictions": {
                    "description": "Associations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Prediction"
                    }
                },
                "score_a": {
                    "description": "Team A final score",
                    "type": "integer"
                },
                "score_b": {
                    "description": "Team B final score",
                    "type": "integer"
                },
                "start_time": {
                    "description": "Scheduled start time",
                    "type": "string"
                },
                "status": {
                    "description": "Current match status",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.MatchStatus"
                        }
                    ]
                },
                "team_a": {
                    "description": "First team name",
                    "type": "string"
                },
                "team_b": {
                    "description": "Second team name",
                    "type": "string"
                },
                "tournament": {
                    "description": "Tournament type",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Tournament"
                        }
                    ]
                },
                "updated_at": {
                    "description": "Record last update timestamp",
                    "type": "string"
                },
                "winner": {
                    "description": "Winner identifier ('A', 'B', or empty)",
                    "type": "string"
                }
            }
        },
        "match.SetResultRequest": {
            "type": "object",
            "properties": {
                "score_a": {
                    "type": "integer",
                    "minimum": 0
                },
                "score_b": {
                    "type": "integer",
                    "minimum": 0
                },
                "winner": {
                    "type": "string",
                    "enum": [
                        "A",
                        "B",
                        ""
                    ]
                }
            }
        },
...
        "match.UpdateMatchRequest": {
            "type": "object",
            "properties": {
                "start_time": {
                    "type": "string"
                },
                "team_a": {
                    "type": "string",
                    "maxLength": 100
                },
                "team_b": {
                    "type": "string",
                    "maxLength": 100
                },
                "tournament": {
                    "$ref": "#/definitions/match.Tournament"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response payload (success only)"
                },
                "error": {
                    "description": "Error details (failure only)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.ErrorInfo"
                        }
                    ]
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "description": "Operation success status",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer \" + JWT令牌进行身份验证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}